python project/kaspi_notifications-main/kaspi_notifications/api/serializers.py
from rest_framework import serializers

from notifications.models import CustomUser


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = CustomUser
        fields = ('id', )
