python project/kaspi_notifications-main/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.4
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
decorator==5.1.1
Django==4.2.9
django-celery-beat==2.7.0
django-filter==24.3
django-timezone-field==7.0
djangorestframework==3.15.2
exceptiongroup==1.2.0
executing==2.1.0
frozenlist==1.4.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
ipython==8.10.0
jedi==0.19.1
Jinja2==3.1.4
kombu==5.3.5
magic-filter==1.0.12
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
multidict==6.0.5
packaging==23.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
prompt-toolkit==3.0.43
propcache==0.2.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.18.0
python-crontab==3.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-telegram-bot==21.0.1
pytz==2024.1
requests==2.31.0
sentry-sdk==2.19.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
